requirements.txt
python-binance
python-dotenv
matplotlib
Pillow
tk